nju/requirements.txt
bs4
pyexecjs